Fixed MILP encoding of hybrid systems
Two issues made the MILP encoding incorrect: the elemental internal
force was being calculated as Kcur * d instead of Kcur * td (this
wasn't an issue in non hybrid code since the auxiliary _int_force
function was being called with td in that case, but in hybrid code
that choice is made when computing the elemental internal force.
Second, Kcur is computed at the end of the loop, which means the
first Kcur (corresponding to deltas[0]) is used twice.